Refactor builtin functions to assembly
