+++ src/shaper/templates.ts
@@ -111,6 +111,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'pow',
         declaration: 'long',
         sentences: [],
+        builtin: 'special',
         name: 'pow'
     },
     {
@@ -118,6 +119,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'mdv',
         declaration: 'long',
         sentences: [],
+        builtin: 'special',
         name: 'mdv'
     },
     {
@@ -125,6 +127,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'powf',
         declaration: 'long',
         sentences: [],
+        builtin: 'special',
         name: 'powf'
     },
     {
@@ -153,6 +156,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'memcopy',
         declaration: 'void',
         sentences: [],
+        builtin: 'internal',
         name: 'memcopy'
     },
     {
@@ -160,6 +164,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'bcftol',
         declaration: 'long',
         sentences: [],
+        builtin: 'internal',
         name: 'bcftol'
     },
     {
@@ -167,6 +172,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'bcltof',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'internal',
         name: 'bcltof'
     },
     {
@@ -174,6 +180,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getNextTx',
         declaration: 'long',
         sentences: [],
+        builtin: 'loop',
         name: 'getNextTx'
     },
     {
@@ -181,6 +188,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getNextTxFromBlockheight',
         declaration: 'long',
         sentences: [],
+        builtin: 'loop',
         name: 'getNextTxFromBlockheight'
     },
     {
@@ -188,6 +196,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getBlockheight',
         declaration: 'long',
         sentences: [],
+        builtin: 'receive',
         name: 'getBlockheight'
     },
     {
@@ -195,6 +204,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCurrentBlockheight',
         declaration: 'long',
         sentences: [],
+        builtin: 'blockchain',
         name: 'getCurrentBlockheight'
     },
     {
@@ -202,6 +212,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getAmount',
         declaration: 'long',
         sentences: [],
+        builtin: 'receive',
         name: 'getAmount'
     },
     {
@@ -209,6 +220,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getSender',
         declaration: 'long',
         sentences: [],
+        builtin: 'receive',
         name: 'getSender'
     },
     {
@@ -216,6 +228,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getType',
         declaration: 'long',
         sentences: [],
+        builtin: 'receive',
         name: 'getType'
     },
     {
@@ -223,6 +236,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCreator',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getCreator'
     },
     {
@@ -230,6 +244,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCreatorOf',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getCreatorOf'
     },
     {
@@ -237,6 +252,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCodeHashOf',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getCodeHashOf'
     },
     {
@@ -244,6 +260,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getWeakRandomNumber',
         declaration: 'long',
         sentences: [],
+        builtin: 'blockchain',
         name: 'getWeakRandomNumber'
     },
     {
@@ -251,6 +268,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getActivationOf',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getActivationOf'
     },
     {
@@ -258,6 +276,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCurrentBalance',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getCurrentBalance'
     },
     {
@@ -265,6 +284,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'readMessage',
         declaration: 'void',
         sentences: [],
+        builtin: 'receive',
         name: 'readMessage'
     },
     {
@@ -272,6 +292,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'readShortMessage',
         declaration: 'void',
         sentences: [],
+        builtin: 'receive',
         name: 'readShortMessage'
     },
     {
@@ -279,6 +300,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendMessage',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendMessage'
     },
     {
@@ -286,6 +308,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendShortMessage',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendShortMessage'
     },
     {
@@ -293,6 +316,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendAmountAndMessage',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendAmountAndMessage'
     },
     {
@@ -300,6 +324,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendAmount',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendAmount'
     },
     {
@@ -307,6 +332,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendBalance',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendBalance'
     },
     {
@@ -314,6 +340,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getMapValue',
         declaration: 'long',
         sentences: [],
+        builtin: 'maps',
         name: 'getMapValue'
     },
     {
@@ -321,6 +348,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getExtMapValue',
         declaration: 'long',
         sentences: [],
+        builtin: 'maps',
         name: 'getExtMapValue'
     },
     {
@@ -328,6 +356,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'setMapValue',
         declaration: 'void',
         sentences: [],
+        builtin: 'maps',
         name: 'setMapValue'
     },
     {
@@ -335,6 +364,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'issueAsset',
         declaration: 'long',
         sentences: [],
+        builtin: 'assets',
         name: 'issueAsset'
     },
     {
@@ -342,6 +372,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'mintAsset',
         declaration: 'void',
         sentences: [],
+        builtin: 'assets',
         name: 'mintAsset'
     },
     {
@@ -349,6 +380,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendQuantity',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendQuantity'
     },
     {
@@ -356,6 +388,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendQuantityAndAmount',
         declaration: 'void',
         sentences: [],
+        builtin: 'fourArgsPlus',
         name: 'sendQuantityAndAmount'
     },
     {
@@ -363,6 +396,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getAssetBalance',
         declaration: 'long',
         sentences: [],
+        builtin: 'contract',
         name: 'getAssetBalance'
     },
     {
@@ -370,6 +404,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'checkSignature',
         declaration: 'long',
         sentences: [],
+        builtin: 'fourArgsPlus',
         name: 'checkSignature'
     },
     {
@@ -377,6 +412,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'distributeToHolders',
         declaration: 'void',
         sentences: [],
+        builtin: 'fourArgsPlus',
         name: 'distributeToHolders'
     },
     {
@@ -384,6 +420,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getAssetHoldersCount',
         declaration: 'long',
         sentences: [],
+        builtin: 'assets',
         name: 'getAssetHoldersCount'
     },
     {
@@ -391,6 +428,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'readAssets',
         declaration: 'void',
         sentences: [],
+        builtin: 'receive',
         name: 'readAssets'
     },
     {
@@ -398,6 +436,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getQuantity',
         declaration: 'long',
         sentences: [],
+        builtin: 'receive',
         name: 'getQuantity'
     },
     {
@@ -405,6 +444,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getAssetCirculating',
         declaration: 'long',
         sentences: [],
+        builtin: 'assets',
         name: 'getAssetCirculating'
     },
 
@@ -414,6 +454,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getAmountFx',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'receive',
         name: 'getAmountFx'
     },
     {
@@ -421,6 +462,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getActivationOfFx',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'contract',
         name: 'getActivationOfFx'
     },
     {
@@ -428,6 +470,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getCurrentBalanceFx',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'contract',
         name: 'getCurrentBalanceFx'
     },
     {
@@ -435,6 +478,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendAmountAndMessageFx',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendAmountAndMessageFx'
     },
     {
@@ -442,6 +486,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendAmountFx',
         declaration: 'void',
         sentences: [],
+        builtin: 'send',
         name: 'sendAmountFx'
     },
     {
@@ -449,6 +494,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getMapValueFx',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'maps',
         name: 'getMapValueFx'
     },
     {
@@ -456,6 +502,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'getExtMapValueFx',
         declaration: 'fixed',
         sentences: [],
+        builtin: 'maps',
         name: 'getExtMapValueFx'
     },
     {
@@ -463,6 +510,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'setMapValueFx',
         declaration: 'void',
         sentences: [],
+        builtin: 'maps',
         name: 'setMapValueFx'
     },
     {
@@ -470,6 +518,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'sendQuantityAndAmountFx',
         declaration: 'void',
         sentences: [],
+        builtin: 'fourArgsPlus',
         name: 'sendQuantityAndAmountFx'
     },
     {
@@ -477,6 +526,7 @@ export const BuiltInTemplate: SC_FUNCTION[] = [
         asmName: 'distributeToHoldersFx',
         declaration: 'void',
         sentences: [],
+        builtin: 'fourArgsPlus',
         name: 'distributeToHoldersFx'
     }
 ]
